+++ plugin.video.oxygen/changelog.txt
@@ -1,8 +1,2 @@
-Version 1.0.1 Initial release
-Version 1.0.2 Added subtitles, metadata, fixed views
-Version 1.0.3 Added user select views
-Version 1.0.4 Website change
-Version 1.0.5 Website change
-Version 1.0.6 cache fix
-Version 1.0.7 website change
+Version 3.0.3 website changes
 Version 3.0.1 Isengard - separate scraper
